+++ src/device/contact.ts
@@ -342,7 +342,7 @@ export class Contact extends deviceBase {
   async registerWebhook() {
     if (this.device.webhook) {
       await this.debugLog('is listening webhook.');
-      this.webhookEventHandler[this.device.deviceId] = async (context: contactSensorWebhookContext) => {
+      this.platform.webhookEventHandler[this.device.deviceId] = async (context: contactSensorWebhookContext) => {
         try {
           await this.debugLog(`received Webhook: ${JSON.stringify(context)}`);
           this.webhookContext = context;
