+++ src/device/humidifier.ts
@@ -373,7 +373,7 @@ export class Humidifier extends deviceBase {
   async registerWebhook() {
     if (this.device.webhook) {
       await this.debugLog('is listening webhook.');
-      this.webhookEventHandler[this.device.deviceId] = async (context: humidifierWebhookContext) => {
+      this.platform.webhookEventHandler[this.device.deviceId] = async (context: humidifierWebhookContext) => {
         try {
           await this.debugLog(`received Webhook: ${JSON.stringify(context)}`);
           this.webhookContext = context;
