+++ src/device/lightstrip.ts
@@ -356,7 +356,7 @@ export class StripLight extends deviceBase {
   async registerWebhook() {
     if (this.device.webhook) {
       await this.debugLog('is listening webhook.');
-      this.webhookEventHandler[this.device.deviceId] = async (context: stripLightWebhookContext) => {
+      this.platform.webhookEventHandler[this.device.deviceId] = async (context: stripLightWebhookContext) => {
         try {
           await this.debugLog(`received Webhook: ${JSON.stringify(context)}`);
           this.webhookContext = context;
