+++ src/device/meterplus.ts
@@ -321,7 +321,7 @@ export class MeterPlus extends deviceBase {
   async registerWebhook() {
     if (this.device.webhook) {
       await this.debugLog('is listening webhook.');
-      this.webhookEventHandler[this.device.deviceId] = async (context: meterPlusWebhookContext) => {
+      this.platform.webhookEventHandler[this.device.deviceId] = async (context: meterPlusWebhookContext) => {
         try {
           await this.debugLog(`received Webhook: ${JSON.stringify(context)}`);
           this.webhookContext = context;
