+++ src/device/robotvacuumcleaner.ts
@@ -304,7 +304,7 @@ export class RobotVacuumCleaner extends deviceBase {
   async registerWebhook() {
     if (this.device.webhook) {
       await this.debugLog('is listening webhook.');
-      this.webhookEventHandler[this.device.deviceId] = async (context: robotVacuumCleanerS1WebhookContext
+      this.platform.webhookEventHandler[this.device.deviceId] = async (context: robotVacuumCleanerS1WebhookContext
         | robotVacuumCleanerS1PlusWebhookContext | floorCleaningRobotS10WebhookContext) => {
         try {
           await this.debugLog(`received Webhook: ${JSON.stringify(context)}`);
